planchet

Struct Client

pub struct Client { /* private fields */ }

The main client for interacting with the Numista API.

Implementations

impl Client

pub async fn get_type( &self, type_id: i64, lang: Option<Language>, ) -> Result<NumistaType>

Gets a single type from the Numista catalogue.

Arguments

type_id - The ID of the type to get.

lang - The language to use for the response.

pub async fn get_issues( &self, type_id: i64, lang: Option<Language>, ) -> Result<Vec<Issue>>

Gets the issues of a type.

Arguments

type_id - The ID of the type to get the issues for.

lang - The language to use for the response.

pub async fn get_prices( &self, type_id: i64, issue_id: i64, currency: Option<&str>, lang: Option<Language>, ) -> Result<PricesResponse>

Gets the prices for an issue.

Arguments

type_id - The ID of the type.

issue_id - The ID of the issue.

currency - The currency to get the prices in.

pub async fn search_types( &self, params: &SearchTypesParams, ) -> Result<SearchTypesResponse>

Searches for types in the Numista catalogue.

Arguments

params - The search parameters.

pub async fn get_issuers( &self, lang: Option<Language>, ) -> Result<IssuersResponse>

Gets the list of issuers.

Arguments

lang - The language to use for the response.

pub async fn get_mints(&self, lang: Option<Language>) -> Result<MintsResponse>

Gets the list of mints.

Arguments

lang - The language to use for the response.

pub async fn get_mint( &self, mint_id: i64, lang: Option<Language>, ) -> Result<MintDetail>

Gets a single mint.

Arguments

mint_id - The ID of the mint to get.

lang - The language to use for the response.

pub async fn get_catalogues(&self) -> Result<CataloguesResponse>

Gets the list of catalogues.

pub async fn get_publication(&self, id: &str) -> Result<Publication>

Gets a single publication.

Arguments

id - The ID of the publication to get.

pub async fn get_user( &self, user_id: i64, lang: Option<Language>, ) -> Result<User>

Gets a user.

Arguments

user_id - The ID of the user to get.

lang - The language to use for the response.

pub async fn get_user_collections( &self, user_id: i64, ) -> Result<CollectionsResponse>

Gets the collections of a user.

Arguments

user_id - The ID of the user to get the collections for.

pub async fn get_collected_items( &self, user_id: i64, params: &GetCollectedItemsParams, ) -> Result<CollectedItemsResponse>

Gets the collected items of a user.

Arguments

user_id - The ID of the user to get the collected items for.

params - The search parameters.

pub async fn add_collected_item( &self, user_id: i64, item: &AddCollectedItem, ) -> Result<CollectedItem>

Adds a collected item to a userโs collection.

Arguments

user_id - The ID of the user to add the collected item to.

item - The item to add.

pub async fn get_collected_item( &self, user_id: i64, item_id: i64, ) -> Result<CollectedItem>

Gets a single collected item from a userโs collection.

Arguments

user_id - The ID of the user.

item_id - The ID of the item to get.

pub async fn edit_collected_item( &self, user_id: i64, item_id: i64, item: &EditCollectedItem, ) -> Result<CollectedItem>

Edits a collected item in a userโs collection.

Arguments

user_id - The ID of the user.

item_id - The ID of the item to edit.

item - The fields to edit.

pub async fn delete_collected_item( &self, user_id: i64, item_id: i64, ) -> Result<()>

Deletes a collected item from a userโs collection.

Arguments

user_id - The ID of the user.

item_id - The ID of the item to delete.

pub async fn get_oauth_token( &self, params: &OAuthTokenParams, ) -> Result<OAuthToken>

Gets an OAuth token.

Arguments

params - The parameters for getting the token.

pub async fn search_by_image( &self, request: &SearchByImageRequest, ) -> Result<SearchByImageResponse>

Searches for types by image.

Arguments

request - The request body.